GRADEME-105: Add intelligence to reset_block task.
For sites with large amounts of data, when the reset block task is run,
the query for finding unmarked quiz questions fails due to a large temp
file being created. The routine has been updated to only update courses
that are active and have active users in the course. And the update is only
run per each individual course. Lastly the query is run in a different
way so that it does not use group by on large tables, which should allow
the process to run on large tables.