Fix issues where learners to be graded quizzes dont appear
This happens when a bunch of learners are regraded. Gradeable items are keyed up on the timemodified for a learners quiz attempt. Regrading a bunch of learners gives them the same timemodified value so when gradeables are added to the array the latest version of the same timemodified quiz attempt overrides the previous version.

1. Update sql for quiz to use timemodified as timemodified and timefinished as timefinished
2. Generate a key that is unique for a learner, course and quiz attempt
3. Use the real time submitted when outputting